refactor: dedupe conversion of historical proposal into UIProposals
